Relocate actix-web service starter to application
1
