+++ application/Cargo.toml
@@ -7,13 +7,17 @@ edition.workspace = true
 
 [dependencies]
 actix-web.workspace = true
+actix-cors.workspace = true
 derive-new.workspace = true
 enum_dispatch.workspace = true
 smart-default.workspace = true
 lazy_static.workspace = true
+log.workspace = true
+env_logger.workspace = true
 
 common_types = { path = "../common/types" }
 common_events = { path = "../common/events" }
 domain = { path = "../domain" }
 in_memory_persistence = { path = "../in_memory_persistence" }
 usecase = { path = "../usecase" }
+rest = {path = "../rest"}
