+++ application/src/main/configuration/mod.rs
@@ -1,3 +1,4 @@
 pub mod application_configuration;
 pub mod persistence_configuration;
+pub mod rest_configuration;
 pub mod use_case_configuration;
